rename: lib -> libs (폴더명 표기법 통일)
